fix(publish-page): print rules must sit below the screen rules they override
The @media print block was placed above the figure sizing rule, .fig-viewport
and .fig-tools. Its selectors tie with theirs on specificity, so source order
decided and every print declaration lost: under print emulation the figure's
max-height stayed 540px, the toolbar stayed display:flex and the scroll box
stayed overflow:auto. Both themes now carry one print block below everything it
overrides, merged with the lightbox hide that was already there.

The guard that read "max-height: none" out of the print block's text could not
have caught this — a tying rule is not evidence about a cascade. It is replaced
by a computed-style measurement under Emulation.setEmulatedMedia: fitted, print
resolves max-height none, .fig-tools display none, .fig-viewport overflow
visible, and the 717x1292 figure renders 1292 tall; zoomed, the box grows past
the svg on paper where it clips and scrolls on screen. A mutation that inserts a
tying rule after the print block is caught.

Also from review:

- The deck's window keydown advanced the slide on Space and preventDefaulted the
  activation away, so Space on a focused zoom button moved the deck and did not
  zoom. Space is now ignored on a button, input, link, editable or anything in
  .fig-tools; arrow keys have no native meaning there and keep navigating. Tested
  with real key events through the input queue, with the body case as the control.
- Zooming below 1x left the box at the fitted height, so 0.5x sat in 540px of
  empty ground. The box now tracks min(fitted, scaled).
- The zoom-in cursor stayed on a zoomed figure although clicking it pans rather
  than expands. .figure.fig-zoomed drops it.
- The zoom factor is rounded to four decimals, so two zoom-outs report 0.64
  rather than 0.6400000000000001.
- The figure cases moved to tests/publish-page/figure-tools.test.ts, registered
  in the product slice; theme-chrome.test.ts drops from 923 to 583 lines. The
  stylesheet rule splitter both files use moved to tests/publish-page/theme-css.ts.

bounded-run: bun test tests/publish-page tests/diagram 782 pass 0 fail;
scripts/preflight --slice product all checks pass, 636 pass 0 fail.

Refs #459
