+++ plugins-cc/agentkit/skills/publish-page/themes/deck.html
@@ -215,16 +215,6 @@
   html[data-theme="light"] svg[role="img"]:not(.edges):not(.d2):not(.drawio) {
     filter: invert(0.93) hue-rotate(180deg);
   }
-  @media print {
-    .figure > svg:not(.edges), .figure > p > svg:not(.edges),
-    .figure > .fig-viewport > svg:not(.edges), .figure > .fig-viewport > p > svg:not(.edges) {
-      -webkit-print-color-adjust: exact; print-color-adjust: exact; max-height: none;
-    }
-    /* A reader can send a zoomed figure to the printer, and the zoom lives in an
-       inline height on the scroll box — which paper has no way to scroll. */
-    .fig-viewport { overflow: visible; height: auto !important; }
-    .fig-tools { display: none; }
-  }
   /* Fit-to-screen cap: a figure is read at a glance, so it is bounded by the
      slide AND by the viewport height, and the toolbar is how a reader gets past
      that. A replaced element with an intrinsic ratio satisfies both caps at
@@ -245,11 +235,13 @@
   }
   .figure p { max-width: none; }
   .fig-viewport { overflow: auto; max-width: 100%; }
-  .fig-viewport.fig-zoomed { cursor: default; }
 
   /* Per-figure toolbar — zoom out/in/reset and full screen. Hover reveals it on
      a pointer device; a touch screen has no hover, so there it always shows. */
   .figure.fig-expandable { cursor: zoom-in; }
+  /* Clicking a zoomed figure pans it rather than expanding it, so the
+     zoom-in cursor would be promising something that no longer happens. */
+  .figure.fig-zoomed { cursor: default; }
   .fig-tools {
     position: absolute; top: 0.55rem; right: 0.7rem; z-index: 2;
     display: flex; gap: 0.25rem; opacity: 0; transition: opacity 0.15s;
@@ -288,7 +280,20 @@
     display: flex; align-items: center; justify-content: center;
   }
   .fig-lightbox-close:hover { border-color: var(--accent); }
-  @media print { .fig-lightbox { display: none; } }
+  /* Print sits below every rule it overrides. These selectors tie on specificity
+     with the screen ones, so source order is the whole cascade: the same block
+     placed above them lost every declaration silently. */
+  @media print {
+    .fig-lightbox { display: none; }
+    .figure > svg:not(.edges), .figure > p > svg:not(.edges),
+    .figure > .fig-viewport > svg:not(.edges), .figure > .fig-viewport > p > svg:not(.edges) {
+      -webkit-print-color-adjust: exact; print-color-adjust: exact; max-height: none;
+    }
+    /* A reader can send a zoomed figure to the printer, and the zoom lives in an
+       inline height on the scroll box — which paper has no way to scroll. */
+    .fig-viewport { overflow: visible; height: auto !important; }
+    .fig-tools { display: none; }
+  }
   pre.mermaid {
     background: var(--card); border: 1px solid var(--line); border-radius: 10px;
     padding: 1rem; overflow-x: auto; text-align: center;
@@ -441,9 +446,18 @@
   }
   prevBtn.addEventListener("click", () => show(i - 1));
   nextBtn.addEventListener("click", () => show(i + 1));
+  const CONTROL = "button, input, select, textarea, a[href], [contenteditable], .fig-tools";
   addEventListener("keydown", (e) => {
     if (document.querySelector(".fig-lightbox:not([hidden])")) return;
-    if (["ArrowRight", "PageDown", " "].includes(e.key)) { e.preventDefault(); show(i + 1); }
+    // Space activates whatever control has focus. Advancing the slide as well
+    // preventDefaults the activation away, so a focused zoom button does nothing
+    // but move the deck out from under the reader. Arrow keys have no such
+    // native meaning on a button, so they keep navigating.
+    const onControl = e.target instanceof Element && e.target.closest(CONTROL);
+    if (["ArrowRight", "PageDown"].includes(e.key) || (e.key === " " && !onControl)) {
+      e.preventDefault();
+      show(i + 1);
+    }
     if (["ArrowLeft", "PageUp"].includes(e.key)) { e.preventDefault(); show(i - 1); }
     if (e.key === "Home") show(0);
     if (e.key === "End") show(slides.length - 1);
@@ -514,7 +528,9 @@
     const svg = zoomTargets(fig)[0];
     if (!svg) return;
     const box = fig.querySelector(".fig-viewport");
-    const next = Math.min(ZOOM_MAX, Math.max(ZOOM_MIN, factor));
+    // Rounded, or repeated division leaves 0.6400000000000001 in the attribute
+    // and the reset comparison below has to guess at a tolerance.
+    const next = Math.min(ZOOM_MAX, Math.max(ZOOM_MIN, Math.round(factor * 1e4) / 1e4));
     // The fitted size is re-measured from a cleared state on every step, so a
     // resized window or a re-rendered figure never scales off a stale baseline.
     if (box) box.style.height = "";
@@ -522,13 +538,13 @@
     const fit = svg.getBoundingClientRect();
     fig.dataset.figZoom = String(next);
     if (next !== 1) {
-      if (box) box.style.height = Math.round(fit.height) + "px";
+      if (box) box.style.height = Math.round(fit.height * Math.min(1, next)) + "px";
       svg.style.maxWidth = "none";
       svg.style.maxHeight = "none";
       svg.style.width = Math.round(fit.width * next) + "px";
       svg.style.height = Math.round(fit.height * next) + "px";
     }
-    box?.classList.toggle("fig-zoomed", next !== 1);
+    fig.classList.toggle("fig-zoomed", next !== 1);
     fig.querySelectorAll(".fig-tool").forEach((button) => {
       const bound = { in: next >= ZOOM_MAX, out: next <= ZOOM_MIN, reset: next === 1 }[button.dataset.figTool];
       if (bound !== undefined) button.disabled = bound;
